Refactor tree_id to u16.

+++ src/lsm/table.zig
@@ -479,7 +479,7 @@ pub fn TableType(
                 cluster: u32,
                 address: u64,
                 snapshot_min: u64,
-                tree_id: u128,
+                tree_id: u16,
             };
 
             pub fn filter_block_finish(builder: *Builder, options: FilterFinishOptions) void {
@@ -529,7 +529,7 @@ pub fn TableType(
                 cluster: u32,
                 address: u64,
                 snapshot_min: u64,
-                tree_id: u128,
+                tree_id: u16,
             };
 
             pub fn index_block_finish(builder: *Builder, options: IndexFinishOptions) TableInfo {
